Relocate xcsv format functions.
Many funtions for the xcsv format were in csv_util.cc
instead of xcsv.cc.  This commit moves the xcsv format
specific functions to xcsv.cc, and leaves the possibly
shared utility functions in csv_util.cc.

xcsv.h is added to communicate shared xcsv style handling
with vecs.cc.
