+++ CMakeLists.txt
@@ -149,6 +149,7 @@ set(HEADERS
   session.h
   shapelib/shapefil.h
   strptime.h
+  xcsv.h
   xmlgeneric.h
   zlib/crc32.h
   zlib/deflate.h
